fix: rename oscillator creator
