+++ packages/UI/synth/index.ts
@@ -1,4 +1,4 @@
 export { createGain } from "./src/create-gain";
-export { createOscillator } from "./src/create-oscilllator";
+export { createOscillator } from "./src/create-oscillator";
 export { play } from "./src/play";
 export { play_note } from "./src/play-note";
